backend/requirements.txt
aniso8601==8.0.0
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dnspython==1.16.0
facebook-sdk==3.1.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SocketIO==4.2.1
gevent==1.4.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
greenlet==0.4.15
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
monotonic==1.5
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pyrsistent==0.15.4
python-engineio==3.9.3
python-gitlab==1.11.0
python-socketio==4.3.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
